# packages/api-contract/openapi.json
{
  "components": {
    "schemas": {
...
      "ArtworkAnalysis": {
        "properties": {
          "findings": {
            "items": {
              "$ref": "#/components/schemas/QualityFinding"
            },
            "type": "array"
          },
          "metrics": {
            "allOf": [
              {
                "$ref": "#/components/schemas/QualityMetrics"
              }
            ],
            "nullable": true
          },
          "status": {
            "enum": [
              "PENDING",
              "PROCESSING",
              "PASSED",
              "NEEDS_RETAKE",
              "FAILED"
            ],
            "type": "string"
          },
          "thresholdVersion": {
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "findings",
          "metrics",
          "status",
          "thresholdVersion"
        ],
        "type": "object"
      },
      "ArtworkAnalysisView": {
        "properties": {
          "analysis": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ArtworkAnalysis"
              }
            ],
            "nullable": true
          },
          "artworkId": {
            "format": "uuid",
            "type": "string"
          },
          "artworkStatus": {
            "enum": [
              "PENDING_UPLOAD",
              "UPLOADED",
              "PROCESSING",
              "READY",
              "FAILED",
              "DELETION_PENDING",
              "DELETED"
            ],
            "type": "string"
          },
          "confirmedCharacter": {
            "nullable": true,
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "analysis",
          "artworkId",
          "artworkStatus",
          "confirmedCharacter",
          "createdAt"
        ],
        "type": "object"
      },
      "ArtworkDeletionView": {
        "properties": {
          "artworkId": {
            "format": "uuid",
            "type": "string"
          },
          "deletionId": {
            "format": "uuid",
            "type": "string"
          },
          "status": {
            "enum": [
              "DELETION_PENDING",
              "DELETED",
              "FAILED"
            ],
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "artworkId",
          "deletionId",
          "status",
          "updatedAt"
        ],
        "type": "object"
      },
...
      "CreateFavoriteGroupRequest": {
        "properties": {
          "name": {
            "maxLength": 40,
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
...
      "CreatePracticeRequest": {
        "properties": {
          "artworkId": {
            "format": "uuid",
            "type": "string"
          },
          "glyphId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "artworkId",
          "glyphId"
        ],
        "type": "object"
      },
...
      "DeleteFavoriteGroupResult": {
        "properties": {
          "deleted": {
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "deleted",
          "id"
        ],
        "type": "object"
      },
      "FavoriteGlyphItem": {
        "properties": {
          "character": {
            "type": "string"
          },
          "favoriteId": {
            "format": "uuid",
            "type": "string"
          },
          "glyph": {
            "$ref": "#/components/schemas/PublishedGlyph"
          },
          "groupId": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "sortOrder": {
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "character",
          "favoriteId",
          "glyph",
          "groupId",
          "sortOrder"
        ],
        "type": "object"
      },
      "FavoriteGroup": {
        "properties": {
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/FavoriteGlyphItem"
            },
            "type": "array"
          },
          "name": {
            "type": "string"
          },
          "sortOrder": {
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "id",
          "items",
          "name",
          "sortOrder"
        ],
        "type": "object"
      },
      "FavoriteGroupResult": {
        "properties": {
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "sortOrder": {
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "sortOrder"
        ],
        "type": "object"
      },
      "FavoriteLibrary": {
        "properties": {
          "groups": {
            "items": {
              "$ref": "#/components/schemas/FavoriteGroup"
            },
            "type": "array"
          },
          "ungrouped": {
            "items": {
              "$ref": "#/components/schemas/FavoriteGlyphItem"
            },
            "type": "array"
          }
        },
        "required": [
          "groups",
          "ungrouped"
        ],
        "type": "object"
      },
      "FavoritePlacementRequest": {
        "properties": {
          "groupId": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
      "FavoriteStateResult": {
        "properties": {
          "favorite": {
            "type": "boolean"
          },
          "glyphId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "favorite",
          "glyphId"
        ],
        "type": "object"
      },
...
      "PracticeAnalysisState": {
        "properties": {
          "failureCode": {
            "nullable": true,
            "type": "string"
          },
          "status": {
            "enum": [
              "PENDING",
              "READY",
              "FAILED"
            ],
            "type": "string"
          }
        },
        "required": [
          "failureCode",
          "status"
        ],
        "type": "object"
      },
      "PracticeAttempt": {
        "properties": {
          "advice": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PracticeAdvice"
              }
            ],
            "nullable": true
          },
          "analysis": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PracticeAnalysisState"
              }
            ],
            "nullable": true
          },
          "artworkId": {
            "format": "uuid",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "imageUrl": {
            "format": "uri",
            "type": "string"
          },
          "sequence": {
            "minimum": 1,
            "type": "integer"
          }
        },
        "required": [
          "advice",
          "analysis",
          "artworkId",
          "createdAt",
          "imageUrl",
          "sequence"
        ],
        "type": "object"
      },
      "PracticeAttemptRequest": {
        "properties": {
          "artworkId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "artworkId"
        ],
        "type": "object"
      },
      "PracticeMaster": {
        "properties": {
          "calligrapherName": {
            "type": "string"
          },
          "glyphId": {
            "format": "uuid",
            "type": "string"
          },
          "imageUrl": {
            "format": "uri",
            "nullable": true,
            "type": "string"
          },
          "workTitle": {
            "type": "string"
          }
        },
        "required": [
          "calligrapherName",
          "glyphId",
          "imageUrl",
          "workTitle"
        ],
        "type": "object"
      },
      "PracticeView": {
        "properties": {
          "attempts": {
            "items": {
              "$ref": "#/components/schemas/PracticeAttempt"
            },
            "type": "array"
          },
          "character": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "master": {
            "$ref": "#/components/schemas/PracticeMaster"
          }
        },
        "required": [
          "attempts",
          "character",
          "createdAt",
          "id",
          "master"
        ],
        "type": "object"
      },
...
      "QualityFinding": {
        "properties": {
          "code": {
            "enum": [
              "BLURRY",
              "INK_TOUCHES_BORDER",
              "LOW_CONTRAST",
              "RESOLUTION_TOO_LOW",
              "TOO_BRIGHT",
              "TOO_DARK"
            ],
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "severity": {
            "enum": [
              "warning",
              "error"
            ],
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "severity"
        ],
        "type": "object"
      },
      "QualityMetrics": {
        "properties": {
          "blur_score": {
            "minimum": 0,
            "type": "number"
          },
          "brightness": {
            "maximum": 255,
            "minimum": 0,
            "type": "number"
          },
          "contrast": {
            "minimum": 0,
            "type": "number"
          },
          "edge_ink_ratio": {
            "maximum": 1,
            "minimum": 0,
            "type": "number"
          },
          "height": {
            "minimum": 1,
            "type": "integer"
          },
          "ink_coverage": {
            "maximum": 1,
            "minimum": 0,
            "type": "number"
          },
          "width": {
            "minimum": 1,
            "type": "integer"
          }
        },
        "required": [
          "blur_score",
          "brightness",
          "contrast",
          "edge_ink_ratio",
          "height",
          "ink_coverage",
          "width"
        ],
        "type": "object"
      },
...
      "ReorderRequest": {
        "properties": {
          "direction": {
            "enum": [
              "UP",
              "DOWN"
            ],
            "type": "string"
          }
        },
        "required": [
          "direction"
        ],
        "type": "object"
      },
      "ReorderResult": {
        "properties": {
          "glyphId": {
            "format": "uuid",
            "type": "string"
          },
          "groupId": {
            "format": "uuid",
            "type": "string"
          },
          "reordered": {
            "type": "boolean"
          }
        },
        "required": [
          "reordered"
        ],
        "type": "object"
      },
...
      "RevokeShareResult": {
        "properties": {
          "revoked": {
            "type": "boolean"
          },
          "shareId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "revoked",
          "shareId"
        ],
        "type": "object"
      },
...
      "ShareResult": {
        "properties": {
          "expiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "url": {
            "format": "uri",
            "type": "string"
          }
        },
        "required": [
          "expiresAt",
          "id",
          "url"
        ],
        "type": "object"
      },
...
      "SwitchPracticeGlyphRequest": {
        "properties": {
          "glyphId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "glyphId"
        ],
        "type": "object"
      },
...
      "UpdateFavoriteGroupRequest": {
        "properties": {
          "name": {
            "maxLength": 40,
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
      "UpdateFavoriteGroupResult": {
        "properties": {
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "移动端、Web 与管理后台共享的业务 API",
    "title": "AI 辅助书法学习 API",
    "version": "1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/artwork-deletions/{deletionId}": {
      "get": {
        "operationId": "PracticeController_getArtworkDeletion",
        "parameters": [
          {
            "in": "path",
            "name": "deletionId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtworkDeletionView"
                }
              }
            },
            "description": "作品物理清理状态。"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "practice"
        ],
        "x-client-contract": true
      }
    },
    "/api/v1/artworks/{artworkId}": {
      "delete": {
        "operationId": "PracticeController_deleteArtwork",
        "parameters": [
          {
            "in": "path",
            "name": "artworkId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtworkDeletionView"
                }
              }
            },
            "description": "作品逻辑删除及物理清理状态。"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "practice"
        ],
        "x-client-contract": true
      },
      "get": {
        "operationId": "AnalysisController_getArtwork",
        "parameters": [
          {
            "in": "path",
            "name": "artworkId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArtworkAnalysisView"
                }
              }
            },
            "description": "作品质检状态。"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "查询本人作品及图片质检状态",
        "tags": [
          "artworks"
        ],
        "x-client-contract": true
      }
    },
...
    "/api/v1/favorite-groups": {
      "post": {
        "operationId": "PracticeController_createFavoriteGroup",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFavoriteGroupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FavoriteGroupResult"
                }
              }
            },
            "description": "收藏分组已创建。"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "practice"
        ],
        "x-client-contract": true
      }
    },
    "/api/v1/favorite-groups/{groupId}": {
      "delete": {
        "operationId": "PracticeController_removeFavoriteGroup",
        "parameters": [
          {
            "in": "path",
            "name": "groupId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteFavoriteGroupResult"
                }
              }
            },
            "description": "收藏分组已删除。"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "practice"
        ],
        "x-client-contract": true
      },
      "patch": {
        "operationId": "PracticeController_updateFavoriteGroup",
        "parameters": [
          {
            "in": "path",
            "name": "groupId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFavoriteGroupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateFavoriteGroupResult"
                }
              }
            },
            "description": "收藏分组已更新。"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "practice"
        ],
        "x-client-contract": true
      }
    },
    "/api/v1/favorite-groups/{groupId}/reorder": {
      "post": {
        "operationId": "PracticeController_reorderFavoriteGroup",
        "parameters": [
          {
            "in": "path",
            "name": "groupId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReorderResult"
                }
              }
            },
            "description": "收藏分组已重新排序。"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "practice"
        ],
        "x-client-contract": true
      }
    },
    "/api/v1/favorites": {
      "get": {
        "operationId": "PracticeController_listFavorites",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FavoriteLibrary"
                }
              }
            },
            "description": "当前用户的收藏字帖。"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "practice"
        ],
        "x-client-contract": true
      }
    },
    "/api/v1/favorites/glyphs/{glyphId}": {
      "delete": {
        "operationId": "PracticeController_unfavorite",
        "parameters": [
          {
            "in": "path",
            "name": "glyphId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FavoriteStateResult"
                }
              }
            },
            "description": "范字收藏已取消。"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "practice"
        ],
        "x-client-contract": true
      },
      "put": {
        "operationId": "PracticeController_favorite",
        "parameters": [
          {
            "in": "path",
            "name": "glyphId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FavoritePlacementRequest"
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FavoriteStateResult"
                }
              }
            },
            "description": "范字已收藏。"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "practice"
        ],
        "x-client-contract": true
      }
    },
    "/api/v1/favorites/glyphs/{glyphId}/reorder": {
      "post": {
        "operationId": "PracticeController_reorderFavorite",
        "parameters": [
          {
            "in": "path",
            "name": "glyphId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReorderResult"
                }
              }
            },
            "description": "收藏范字已重新排序。"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "practice"
        ],
        "x-client-contract": true
      }
    },
...
    "/api/v1/practices": {
      "get": {
        "operationId": "PracticeController_listPractices",
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PracticeView"
                  },
                  "type": "array"
                }
              }
            },
            "description": "当前用户的练习历史。"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "practice"
        ],
        "x-client-contract": true
      },
      "post": {
        "operationId": "PracticeController_createPractice",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePracticeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PracticeView"
                }
              }
            },
            "description": "练习会话已创建。"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "以本人作品和已发布同字范字创建练习记录",
        "tags": [
          "practice"
        ],
        "x-client-contract": true
      }
    },
    "/api/v1/practices/{sessionId}": {
      "get": {
        "operationId": "PracticeController_getPractice",
        "parameters": [
          {
            "in": "path",
            "name": "sessionId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PracticeView"
                }
              }
            },
            "description": "练习会话详情。"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "practice"
        ],
        "x-client-contract": true
      }
    },
    "/api/v1/practices/{sessionId}/attempts": {
      "post": {
        "operationId": "PracticeController_addAttempt",
        "parameters": [
          {
            "in": "path",
            "name": "sessionId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PracticeAttemptRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PracticeView"
                }
              }
            },
            "description": "再次练习已加入会话。"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "practice"
        ],
        "x-client-contract": true
      }
    },
    "/api/v1/practices/{sessionId}/glyph": {
      "patch": {
        "operationId": "PracticeController_switchGlyph",
        "parameters": [
          {
            "in": "path",
            "name": "sessionId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwitchPracticeGlyphRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PracticeView"
                }
              }
            },
            "description": "参考范字已切换。"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "切换练习会话的参考范字",
        "tags": [
          "practice"
        ],
        "x-client-contract": true
      }
    },
    "/api/v1/practices/{sessionId}/shares": {
      "post": {
        "operationId": "PracticeController_createShare",
        "parameters": [
          {
            "in": "path",
            "name": "sessionId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareResult"
                }
              }
            },
            "description": "私密练习分享已创建。"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "practice"
        ],
        "x-client-contract": true
      }
    },
...
    "/api/v1/shares/{shareId}": {
      "delete": {
        "operationId": "PracticeController_revokeShare",
        "parameters": [
          {
            "in": "path",
            "name": "shareId",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevokeShareResult"
                }
              }
            },
            "description": "分享已撤销。"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "practice"
        ],
        "x-client-contract": true
      }
    },
    "/api/v1/shares/{token}": {
      "get": {
        "operationId": "PublicShareController_getShare",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9_-]{40,60}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PracticeView"
                }
              }
            },
            "description": "公开练习分享。"
          }
        },
        "summary": "查看未撤销且未过期的公开练习分享",
        "tags": [
          "shares"
        ],
        "x-client-contract": true
      }
    },
...
  },
  "servers": [],
...
}